added scss for filter button and panel items. added clear button to clear date values
