+++ geograpix/src/scss/index.scss
@@ -23,6 +23,7 @@ code {
 @import './marker';
 @import './picturePopup';
 @import './profile';
+@import './filter';
 
 .ir {
   color: transparent;
